+++ common-libs/common/src/index.ts
@@ -15,11 +15,11 @@ if (typeof process.version === 'undefined') {
 }
 
 export { JwtService } from '@affinidi/tools-common'
-
 export { Affinity, Affinity as Affinidi } from './Affinity'
 
 export {
   KeyVault,
+  EncryptionService,
   LocalKeyVault,
   DidDocumentService,
   DigestService,