+++ common-libs/common/CHANGELOG.md
@@ -1,3 +1,7 @@
+# release 2.9.1 (2022-11-29)
+* add security fix for signIn passwordless flow
+# release 2.9.0 (2022-11-28)
+* updated `internal-api-clients`
 # release 2.8.0 (2022-11-14)
 * updated `url-resolver` with internal URLs for new dev environment
 # release 2.7.0 (2022-11-10)