+++ sdk/core/CHANGELOG.md
@@ -1,3 +1,5 @@
+# release 7.8.0 (2022-11-29)
+* add security fix for signIn passwordless flow
 # release 7.7.0 (2022-11-29)
 feat: add getInfo to NetworkMemberWithCognito
 # release 7.6.0 (2022-11-22)